+++ kernel/src/interpreter/messages.ts
@@ -6,15 +6,36 @@ import {
   ImagePart,
   TextPart,
 } from 'ai';
-import {
-  ActionOutput,
-  FileInput,
-  Interaction,
-  TextOutput,
-} from './interactions';
-import { encodeActionHandle } from '../runtime/actions';
+import { ActionDirective } from '../runtime/actions';
+import { ProcessContainer } from '../runtime/processes';
 
-export type Message =
+export type KernelMessage = InputMessage | ResponseMessage | ActionMessage;
+
+export interface InputMessage {
+  type: 'input';
+  text?: string;
+  files?: FileInput[];
+}
+
+export interface FileInput {
+  data: Uint8Array;
+  filename?: string;
+  mimeType?: string;
+}
+
+export interface ResponseMessage {
+  type: 'response';
+  content: string;
+}
+
+export interface ActionMessage {
+  type: 'action';
+  directive: ActionDirective;
+  process?: ProcessContainer;
+  content?: any;
+}
+
+export type ModelMessage =
   | CoreSystemMessage
   | CoreUserMessage
   | CoreAssistantMessage;
@@ -23,85 +44,86 @@ export function createUserMessage(content: string) {
   return {
     role: 'user',
     content,
-  } as Message;
+  } as ModelMessage;
 }
 
 export function createAssistantMessage(content: string) {
   return {
     role: 'assistant',
     content,
-  } as CoreAssistantMessage;
+  } as ModelMessage;
 }
 
-export function createMessages(
-  interactions: Interaction[],
-  ...prompts: string[] | undefined
-): Message[] {
-  let messages: Message[] = [];
-  for (let interaction of interactions) {
-    if (interaction.input.text)
-      messages.push(createUserMessage(interaction.input.text));
+// export function parseMessages(messages: KernelMessage[]): ModelMessage[] {
+//   let modelMessages: ModelMessage[] = [];
 
-    if (interaction.input.files?.length) {
-      const parts: Array<TextPart | ImagePart | FilePart> =
-        interaction.input.files.map(fileMessage);
+//   for (const message of messages) {
+//     if (message.type === 'input') {
+//       if (message.text) modelMessages.push({
+//         role: 'user',
+//         content: message.text
+//       });
 
-      messages.push({
-        role: 'user',
-        content: parts,
-      });
-    }
+//       if (interaction.input.files?.length) {
+//         const parts: Array<TextPart | ImagePart | FilePart> =
+//           interaction.input.files.map(fileMessage);
 
-    if (!interaction.outputs) continue;
+//         messages.push({
+//           role: 'user',
+//           content: parts,
+//         });
+//       }
 
-    for (let output of interaction.outputs) {
-      if (output.type === 'text') {
-        const textOutput = output as TextOutput;
-        messages.push(createAssistantMessage(textOutput.content));
-      } else if (output.type === 'action') {
-        const { directive, process, content } = output as ActionOutput;
-        const actionUri = encodeActionHandle(directive.uri, directive.actionId);
-        let messageStr = `Action called: ${actionUri}.\n`;
-        const outputStr = JSON.stringify(
-          process ? process.describe() : content
-        );
-        messageStr += `Output: ${outputStr}`;
-        console.log(messageStr);
-        messages.push(createAssistantMessage(messageStr));
-      }
-    }
-  }
+//       if (!interaction.outputs) continue;
 
-  if (prompts) {
-    for (const prompt of prompts) {
-      messages.push({
-        role: 'user',
-        content: prompt,
-      });
-    }
-  }
+//       for (let output of interaction.outputs) {
+//         if (output.type === 'text') {
+//           const textOutput = output as TextOutput;
+//           messages.push(createAssistantMessage(textOutput.content));
+//         } else if (output.type === 'action') {
+//           const { directive, process, content } = output as ActionOutput;
+//           const actionUri = encodeActionHandle(directive.uri, directive.actionId);
+//           let messageStr = `Action called: ${actionUri}.\n`;
+//           const outputStr = JSON.stringify(
+//             process ? process.describe() : content
+//           );
+//           messageStr += `Output: ${outputStr}`;
+//           console.log(messageStr);
+//           messages.push(createAssistantMessage(messageStr));
+//         }
+//       }
+//     }
 
-  return messages;
-}
+//     if (prompts) {
+//       for (const prompt of prompts) {
+//         messages.push({
+//           role: 'user',
+//           content: prompt,
+//         });
+//       }
+//     }
 
-export function fileMessage(file: FileInput): TextPart | ImagePart | FilePart {
-  if (file.mimeType.startsWith('text/') || file.mimeType === 'application/json')
-    return {
-      type: 'text',
-      text: new TextDecoder().decode(file.data),
-    };
+//     return messages;
+//   }
 
-  if (file.mimeType.startsWith('image/'))
-    return {
-      type: 'image',
-      image: file.data,
-      mimeType: file.mimeType,
-    };
+//   export function fileMessage(file: FileInput): TextPart | ImagePart | FilePart {
+//     if (file.mimeType.startsWith('text/') || file.mimeType === 'application/json')
+//       return {
+//         type: 'text',
+//         text: new TextDecoder().decode(file.data),
+//       };
 
-  return {
-    type: 'file',
-    data: file.data,
-    filename: file.filename,
-    mimeType: file.mimeType,
-  };
-}
+//     if (file.mimeType.startsWith('image/'))
+//       return {
+//         type: 'image',
+//         image: file.data,
+//         mimeType: file.mimeType,
+//       };
+
+//     return {
+//       type: 'file',
+//       data: file.data,
+//       filename: file.filename,
+//       mimeType: file.mimeType,
+//     };
+//   }
